requirements.txt
requests==2.28.1
mysql-connector-python==8.0.29